requirements.txt
argparse==1.4.0
jsonschema=4.4.0
numpy==1.21.5
Pillow==9.0.1
pytorch==1.12.0
PyYAML==5.4.1
torchmetrics==0.9.2
torchvision==0.13.0